hackthon/requirements.txt
transformers==4.30.2
torch==2.0.1
datasets==2.12.0
accelerate==0.20.3
flask==2.3.3
python-dotenv==1.0.0
google-generativeai==0.3.1 